requirements.txt
chromadb
requests
tqdm
pypdf
python-docx
pdfminer.six
